Move ShadowRoot from ElementRareData to Element https://bugs.webkit.org/show_bug.cgi?id=296897
Reviewed by NOBODY (OOPS!).

There are many elements with UA shadow roots. Move this pointer from ElementRareData to Element
to eliminate the overhead of allocating ElementRareData for those elements.

* Source/WebCore/dom/Element.cpp:
(WebCore::Element::addShadowRoot):
(WebCore::Element::removeShadowRootSlow):
* Source/WebCore/dom/Element.h:
(WebCore::Element::shadowRoot const):
* Source/WebCore/dom/ElementInlines.h:
* Source/WebCore/dom/ElementRareData.cpp:
* Source/WebCore/dom/ElementRareData.h:
(WebCore::ElementRareData::useTypes const):
(WebCore::ElementRareData::~ElementRareData):
(WebCore::ElementRareData::clearShadowRoot): Deleted.
(WebCore::ElementRareData::shadowRoot const): Deleted.
(WebCore::ElementRareData::setShadowRoot): Deleted.
(WebCore::Node::shadowRoot const): Deleted.
(WebCore::Element::shadowRoot const): Deleted.
(WebCore::Node::protectedShadowRoot const): Deleted.
(WebCore::Element::removeShadowRoot): Deleted.
* Source/WebCore/dom/Node.h:
* Source/WebCore/dom/ShadowRoot.h:
(WebCore::Node::shadowRoot const):
(WebCore::Node::protectedShadowRoot const):
(WebCore::Element::removeShadowRoot):
